requirements/testing.txt
# PyTest for running the tests.
pytest==2.6.4
pytest-django==2.8.0
pytest-cov==1.6